+++ plugins/source/aws/client/resolvers_test.go
@@ -153,7 +153,7 @@ var jsonBytes = []byte(jsonString)
 func TestResolveStringJson(t *testing.T) {
 	cases := []struct {
 		InputItem    interface{}
-		ExpectedData map[string]interface{}
+		ExpectedData interface{}
 		Path         string
 	}{
 		{
@@ -165,9 +165,23 @@ func TestResolveStringJson(t *testing.T) {
 		},
 		{
 			InputItem: struct {
+				Json string
+			}{Json: ""},
+			ExpectedData: nil,
+			Path:         "Json",
+		},
+		{
+			InputItem: struct {
 				Json *string
 			}{Json: &jsonString},
 			ExpectedData: map[string]interface{}{"k1": "v1"},
+			Path:         "Json",
+		},
+		{
+			InputItem: struct {
+				Json *string
+			}{Json: nil},
+			ExpectedData: nil,
 			Path:         "Json",
 		},
 		{
@@ -182,6 +196,20 @@ func TestResolveStringJson(t *testing.T) {
 				Json *[]byte
 			}{Json: &jsonBytes},
 			ExpectedData: map[string]interface{}{"k1": "v1"},
+			Path:         "Json",
+		},
+		{
+			InputItem: struct {
+				Json *[]byte
+			}{Json: nil},
+			ExpectedData: nil,
+			Path:         "Json",
+		},
+		{
+			InputItem: struct {
+				Json []byte
+			}{Json: []byte{}},
+			ExpectedData: nil,
 			Path:         "Json",
 		},
 	}