+++ plugins/source/aws/codegen/recipes/base.go
@@ -59,6 +59,7 @@ var defaultRegionalColumns = []codegen.ColumnDefinition{
 func awsNameTransformer(f reflect.StructField) (string, error) {
 	c := caser.New(caser.WithCustomInitialisms(map[string]bool{
 		"EC2": true,
+		"VPC": true,
 	}))
 	return c.ToSnake(f.Name), nil
 }