+++ plugins/source/aws/resources/services/route53/models/route53.go
@@ -1,0 +1,10 @@
+package models
+
+import "github.com/aws/aws-sdk-go-v2/service/route53/types"
+
+type Route53HostedZoneWrapper struct {
+	types.HostedZone
+	Tags            map[string]string
+	DelegationSetId *string
+	VPCs            []types.VPC
+}